nitinat/requirements.txt
black>=22.1
flake8>=4.0
invoke>=1.7
isort>=5.10
kaleido>=0.2
pandas>=1.4
plotly>=5.6
pydantic>=1.9.0
pyinvokedepends>=1.0
pytest>=7.1
pyyaml>=5.4
requests>=2.27
statsmodels>=0.13
